"Learning hub" view
This is (the start of) an implementation of
https://forums.online-go.com/t/accessible-teaching/11039

The idea is to use a stripped down "Group" view so that we can drive the
content of the learning hub page (news, teachers) from existing
infrastructure.

We'll need to create an invite-only, private "OGS Teachers" group on the
main site to drive the page from, and presumably a `global-teaching`
channel to link with from `EmbeddedChat`.

Once we have this hashed out, next we'd like to create an interactive
tutorial that we'll link to from this page along with a player's
progress - that'll be more involved though.

There are some problems I know about with this initial stab:

*   `PaginatedTable` may not be the right thing to use for the list of
    teachers, given that we'd like to sort the whole group by online
    status, and keep it refreshed.
*   The "Resources" section is anaemic.

Any advice appreciated.

A screenshot running against the Beta server:

*   http://cian.furiousthinking.org/picture/ogs-learn-1.png