Fix middleclick board opening issues on Chrome
Apparently not receiving click events on Chrome for middle clicks is now
intended behavior. As a result, we have to start using onMouseUp to
capture them in most cases, though we also need to use onClick for other
cases still as iOS safari doesn't always seem to work with onMouseUp as
desired.

Fixes #174. Probably breaks other stuff, we'll find out.