Fix clock state reset issue
I'm pretty sure this fixes #354

Problem was we were setting two timers for updateClock because of an
update loop stemming from a reaction to our emitted event (which would
finish evaluating before the assignment of our setTimeout), so after
awhile that method could be called with an old clock object and reset
things in a funky way.

This patch also includes a fix for how often to call the clock update
function, we were stomping a common timer value so it wouldn't always
update right on the second as we desire.