Cause the Player component to use the new Player type. The Player, PlayerDetails and PlayerIcon components now all inherit from an abstract PlayerComponent that handles the component's iteraction with the player_cache. A PlayerComponent has two standard properties -- user: Player | number and using_cache?: boolean.
The latter replaces dontUpdateCache from the Player component, but with slightly different semantics.
Whereas dontUpdateCache only caused the component to not update the cache, the new using_cache property also causes the component's state to follow changes to the player_cache. If the user is specified by id, then using_cache is always assumed to be true. If using_cache is false, the the user will never change, even if the player_cache is updated and some properties of the player have their default values.