Review Tool: Quick Jump to Boardstate #563
Shift+clicking seeks a move played on that point. If the current move is
a sub branch, it will be explored before checking the main branching
point, followed by any other subbranches of that point. Then the trunk
after the main branching point is checked followed by the trunk before
it. If the current move is a trunk move, all subbranches of that move
will be searched before returning to the trunk as above.