Updated automatch to use SFXManager and bug fix related to new game notifications
The SFXManager deals with things like preloading, ogg/mp3/wav support,
and picking the right cdn release base.

The sound play code was moved to the automatch manager instead of the
notifications, as sticking that code in the notifications would cause
the sound to be played on refresh whenever there were pending new game
notifications lingering.