Return a lowercase string from durationString.
Most, if not all, calls to durationString are followed by toLowerCase().  This causes difficulty in German translation because nouns are always capitalized regardless of location in the sentence.  This commit allows translators to control the capitalization of the time words "week", "day", "hour", "minute" and "second".

In addition, the internals of durationString() have been updated to allow for re-ordering of the units, as might be the case in right-to-left languages.  This puts a higher workload on the translators, but gives more flexibility.