ngettext: Don't use plural translation for singular counts and vice versa.
ngettext was a little too smart, causing "1 day" to be translated to "1 days" in durationString() because "%s days" was in the catalog, but not "%s day".  Since there is no English project at translate.online-go.com, there was no way to fix this besides changing "%s day" to "1 day", which causes potential issues with interpolate().