Preserve type when using dup()
This commit changes the siganture of dup() from `dup(obj: any) : any` to `dup<T>(obj: T) : T`.  Since we know the result of dup should be a copy of obj, it makes sense that the return type should be the same as the type of obj.

This may have uncovered a bug in PuzzleEditing where nonexistent members of PuzzleCongig were referenced.