Better default theme based on OS choice
We're now setting what the *default* value of the theme should be
based on the OS theme choice instead of setting the value permenantly
based on the first view. This should also behave better for setting the
default theme for your devices using the new remote storage mechanism,
but still allowing local overrides.