Per-user keyed player notes
This ensures that when you log in as a different user, or are not logged
in, you don't see notes from another account.

While there's a loose idea that should be firmed up regarding keying all
data and preferences with the user id, I have a nagging suscpicion that
we'll want still have these keyed to make our lives easier when it comes
to transitioning data and preferences to those keyed values, and dealing
with what do we do when we change accounts and transition from being
logged out to logged in and vise versa.

This PR stacks on top of #1462