Update PlayerAutocomplete to hook style and fix overzealous complete triggering
Prior to this update PlayerAutocomplete would fire constantly if there
was a player that matched, this is inconvenient when you're using the
input field for a heavy query (such as searching for player games on the
player profile page). Now the behavior is such that we only trigger when
a user selects a user explicitly, presses enter, or blurs out of the
input field.

I also ported this to be a React hook and added a lot of typings.