Have eslint run our prettier checks
Having eslint run our prettier checks gets us some prettier output in
our gulp build loop window so we know when prettier is failing before a
commit.  The gulp-prettier plugin seemed to do the same but with less
information (no line numbers for instance, although I assume I just
didn't have it configured right), and overall it was slower, probably
because it had to run a full prettier check after a full eslint check,
and either eslint's caching is applied to the prettier checks when
eslint does it, or it's run in parallel, or both, I'm not sure - but
this seemed to be the better option, but I'm open to being corrected.