Protect against unsupported up/down branch navigation in conditional move planner
This feature was never implemented properly, and we can get into broken
states when there are analysis branches that already exist while we are
in the conditional move planner.

I think we want to revisit how the conditional move planner works
someday and perhaps integrate it in with the analysis mode branch
explorer, at that time hopefully we can support sibling branch
navigation.