Updated image resizer code to use webp and have a file size limit
This is a workaround to fix #2406 where a chrome behavior (I say bug)
causes a fetch error to occur if our resized files are larger than
64KiB. All resources I've found on the topic do not mention any such
limit and it's unclear to me why this limit is being imposed, so it
may be a temporary issue in Chrome or it may be something else that I'm
not aware of. Regardless, the change to webp largely resolves the issue,
and scaling images down if they are "large" is a fine solution too I
think, so I think the resulting workaround is a desirable behavior
regardless.