Skip empty transmitting in line summaries when paused
f2ebda37ecbe and b2f2565eeae8 already tightened up the display of line
summaries for games that are NOT paused by removing the `<span>` for
`transmitting`.

This follow-up also removes the span when the game IS paused. Otherwise, a
blank line is added on really small screens (e.g., mobile), when the empty
transmitting `<span>` and the word "Paused" don't fit on one line.