Declare GameType.pause_control as optional
Fix the declaration of GameType.pause_control, added in
b90b04a50322a61cee94c3b07f209973c5354411. It's missing when the game has never
been paused.

The code using it below already treated it as optional (and I tested that
scenario), but it's still worth cleaning up the declaration for clarity.