Sort clocks correctly that aren't going to expire
Expose the `JGOFTimeControl` field in `GameType.json`, and use that to sort
clocks that aren't going to expire by computing the time remaining. This
includes:

- Sorting by the clock of the player whose turn it is NOT.
- Sorting by either clock in a paused game.

Note: 7ac13061552abd57827802f19bdde54c098a8308 fixed paused games to not be
intermingled with un-paused games, but was still sorting by an ill-defined
expiration time. This computes the time left on the clock and sorts by that
instead.