Switch GameList to sort based on JGOFClock
Change GameList over to sorting based on the JGOFClock instead of AdHocClock.
Computing remaining time is simpler and easier to get right, especially for
paused games.

The downside is that we don't initially have a JGOFClock available.  Here's
how that's handled:

- New callback to `GobanLineSummary`, called the first time a `JGOFClock` is
  available.
- New field `force_render` in `GameList`'s state, which gets bumped by the
  callback.  When `render()` gets called again, the sort will be reevaluated.

Note: `force_render` is named generically because there might be other
(non-clock) reasons to force `GameList.render()` to be called in the future.

Fixes #2451.