Fix file uploads on chrome/safari
This "keepalive" flag was not what I thought it was, it is for keeping a
request going after a tab or window is closed. Apparently when this is
set and I assume too big of a POST, such as a file upload, is requested
chrome and safari simply throw a rather unhelpful error message.