Neutralize humble_rating
Update `humble_rating` to return the `rating` parameter directly, rather than
modifying it using the deviation, effectively neutralizing it.

After this lands, we should follow up by deleting the function entirely, but
this commit as-written is much less likely to develop conflicts while we're
waiting to land it and/or if we have to revert it.