Consistently select game state between 'demo' and 'challenge.game'
The board settings for a challenge/demo modal are stored in three places in the
state object:

- `state.conf`
- `state.demo` (demo boards)
- `state.challenge.game` (others)

Now, most accessors/modifiers of game state select between `challenge.game` and
`demo` using either `gameState()`, `gameStateOf()`, or `gameStateName()`.

Previously, some of the demo settings were stored temporarily in
`challenge.game` and copied over to `demo`, while others were already split.

This clean-up means that:

- The saved settings in "demo" and "challenge.game" are now independent.
  Previously, some were independent and others were not.
    - "Create Demo" remembers the last demo you made.
    - "Create Challenge" remembers the last challenge you made.
- Rendering code can (still) be easily shared.

With this in place, it will be straightforward to allow users to select
handicap and komi for demo boards.

Relates to #2501.