Only show custom komi in challenge descriptions
The `komi` field is inaccurate (and ignored) when `komi_auto !== "custom"`.  It
might be nice to show the automatic komi, but right now that would be hard to
do so consistently, since for small boards it depends on handicap tables that
are server side.

For now, just stop showing komi at all unless `komi_auto === "custom"`.

Fixes #2527.