Allow groups to turn on handicaps for their ladders
Add frontend support for handicaps in group ladders:

- Enable configuration on the group page.
- Show the setting on the ladder page.
- Turn on handicaps by default when creating new tournaments.

One slightly awkward thing is that the group page canonicalizes the type as a
`number` (0 or -1) and tournament page uses a `string` ("0" or "-1"), depending
on what the backend expects.  If you'd rather have consistency here, I can
change the backend to expect a string for the group page and convert to a
number in the backend.

Fixes #2553