Move rules out of the black player card
In the black player card, the rules codes can be mistaken for initials of the
black player.  Move them out, just below the player cards, into an area for
condensed game information with small font.

- Rules, aligned right.
- Ranked games, "Ranked", aligned left.

This is an option for addressing [feedback from the
forums](https://forums.online-go.com/t/the-ruleset-of-the-game-should-be-easier-to-see/39165).