WIP: Allow tournaments to be cloned
Allow tournaments to be cloned by adding routes:

- `/tournament/clone/7` clones tournament 7 without a group
- `/tournament/clone/7/37` clones tournament 7 into group 37

This works if you manually navigate to the URLs.  It behaves like a new
tournament, loads tournament 7, and copies everything in.

But, the shiny new "Clone Tournament" button doesn't work.  We need a fresh
instance of the `<Tournament/>` component, but it's reusing the existing one.

I'm not sure what the best approach is here.