Add `/reload` route to fix the default view for social logins
Before this commit: when social logins succeed after Register or SignIn views,
they currently show the content from the ObserveGames view instead of from the
Overview view. This is because `config.user.anonymous` is true when the Default
component is loaded.

Username/password signins don't have this problem because they force a reload
by explicitly setting `window.location.pathname`.

This is a speculative fix (I can't test the OAuth stuff), adding a
`/reload` route that uses the same trick for social logins.